v0.3.0
Working on a lot of refactoring to leverage vllm instead of the old transformers-based approach. This new approach has much better throughput, although there are a LOT of quality-of-life things (as well as major exceptions that can crop up) that need to be handled before releasing an updated version of the package.